requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
ultralytics>=8.2.0
Pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
